added home page and will addd functionality soon
